Studio: Image objects no longer copy to Direct Buffers.
Several contstructors of DefaultImage were still using Direct Buffers.
These now have been changed to use normal, Java-Buffers instead.
The utility class DirectBuffers was renamed to BufferTools and moved
to the same package as DefaultImage. Changes to the memory model used
should be easy to make in the future.
