+++ mmstudio/src/main/java/org/micromanager/data/internal/DefaultImage.java
@@ -30,7 +30,6 @@ import java.nio.Buffer;
 import java.nio.ByteBuffer;
 import java.nio.ShortBuffer;
 
-import com.jogamp.common.nio.Buffers;
 import mmcorej.TaggedImage;
 import mmcorej.org.json.JSONException;
 import mmcorej.org.json.JSONObject;
@@ -41,19 +40,17 @@ import org.micromanager.data.Coords;
 import org.micromanager.data.Image;
 import org.micromanager.data.Metadata;
 import org.micromanager.internal.propertymap.NonPropertyMapJSONFormats;
-import org.micromanager.internal.utils.DirectBuffers;
 import org.micromanager.internal.utils.imageanalysis.ImageUtils;
 import org.micromanager.internal.utils.ReportingUtils;
 
 /**
  * This class represents a single image from a single camera. It contains
  * the image pixel data, metadata (in the form of a Metadata instance), and
- * the image's index as part of a larger dataset (in the form of an
- * Coords instance).
+ * the image's index as part of a larger dataset (in the form of a Coords instance).
  *
- * For efficiency during high-speed acquisitions, we store the image data in
- * a ByteBuffer or ShortBuffer. However, ImageJ wants to work with image data
- * in the form of byte[], short[], or int[] arrays (depending on pixel type).
+ * We store the image data in a ByteBuffer or ShortBuffer. However, ImageJ
+ * wants to work with image data in the form of byte[], short[], or
+ * int[] arrays (depending on pixel type).
  * getRawPixels(), the method exposed in the Image interface to access pixel
  * data, returns an ImageJ-style array, while getPixelBuffer (which is not
  * exposed in the API) returns the raw buffer.
@@ -72,11 +69,13 @@ public final class DefaultImage implements Image {
 
    /**
     * Generate a DefaultImage from a TaggedImage. Note that this method will
-    * result in Micro-Manager assuming that the image data was generated by
-    * the scope that it is currently running (i.e. this is a bad method to use
+    * result in Micro-Manager assuming the image data came from
+    * the microscope that is currently running (i.e. this is a bad method to use
     * for loading saved images). If you want to avoid that, then you need to
     * manually reconstruct the Metadata for the TaggedImage and use the
     * constructor that this method calls.
+    *
+    * PixelData from the TaggedImage will be used directly, i.e., they are not copied.
     * @param tagged A TaggedImage to base the Image on.
     */
    public DefaultImage(TaggedImage tagged) throws IllegalArgumentException {
@@ -84,8 +83,16 @@ public final class DefaultImage implements Image {
    }
 
    /**
-    * As above but allows either or both of the image coords and metadata to be
-    * overridden.
+    * Generate a DefaultImage from a TaggedImage using the
+    * supplied coords and metadata.
+    *
+    * PixelData from the TaggedImage will be used directly, i.e., they are not copied.
+    *
+    * @param tagged A TaggedImage to base the Image on. Pixels are not copied.
+    * @param coords Coords to be used for this new Image. When null, tags in input image
+    *               are interpreted to best guess the desired Coord.
+    * @param metadata Metadata to be used for this new Image. When null, tags from input image
+    *                 are used to construct metadata.
     */
    public DefaultImage(TaggedImage tagged, Coords coords, Metadata metadata)
          throws IllegalArgumentException {
@@ -128,12 +135,7 @@ public final class DefaultImage implements Image {
 
       metadata_ = (DefaultMetadata) metadata;
       coords_ = coords;
-/*
-      rawPixels_ = DirectBuffers.bufferFromArray(tagged.pix);
-      if (rawPixels_.capacity() == 0) {
-         throw new IllegalArgumentException("Pixel data has length 0");
-      }
-*/
+
       pixelWidth_ = formatPmap.getInteger(PropertyKey.WIDTH.key(), 0);
       pixelHeight_ = formatPmap.getInteger(PropertyKey.HEIGHT.key(), 0);
       if (pixelWidth_ <= 0 || pixelHeight_ <= 0) {
@@ -145,86 +147,91 @@ public final class DefaultImage implements Image {
          throw new IllegalArgumentException("Missing pixel type");
       }
 
-      switch (pixelType_.getBytesPerComponent()) {
-         case 1:
-            rawPixels_ = ByteBuffer.wrap((byte[]) tagged.pix);
-            if (!(rawPixels_ instanceof ByteBuffer)) {
-               throw new IllegalArgumentException("Array type doesn't match pixel type");
-            }
-            break;
-         case 2:
-            rawPixels_ = ShortBuffer.wrap((short[]) tagged.pix);
-            if (!(rawPixels_ instanceof ShortBuffer)) {
-               throw new IllegalArgumentException("Array type doesn't match pixel type");
-            }
-            break;
-         default:
-            throw new AssertionError("Unimplemented pixel component size");
-      }
+      rawPixels_ = BufferTools.wrapArray(tagged.pix, pixelType_.getBytesPerComponent());
+
       if (rawPixels_.capacity() == 0) {
          throw new IllegalArgumentException("Pixel data has length 0");
       }
    }
 
+   /**
+    * Generates a DefaultImage from pixels, image info in a PropertyMap and the
+    * supplied coords and metadata.
+    *
+    * Input pixels will be used directly (i.e., they are not copied).
+    *
+    * @param pixels Image pixels.  Should be a Java array of bytes or shorts (not null).
+    * @param format PropertyMap specifying image width, height, and pixelType (not null).
+    * @param coords Coords to be used for this new image (can be null).
+    * @param metadata Metadata to be used this new image (can be null).
+    * @throws IllegalArgumentException
+    */
    public DefaultImage(Object pixels, PropertyMap format, Coords coords,
          Metadata metadata) throws IllegalArgumentException {
-      Preconditions.checkNotNull(pixels);
-      Preconditions.checkNotNull(format);
-      metadata_ = metadata == null ? new DefaultMetadata.Builder().build() :
-            (DefaultMetadata) metadata;
-      coords_ = coords == null ? Coordinates.builder().build() : coords;
-      rawPixels_ = DirectBuffers.bufferFromArray(pixels);
-      pixelWidth_ = format.getInteger(PropertyKey.WIDTH.key(), 0);
-      pixelHeight_ = format.getInteger(PropertyKey.HEIGHT.key(), 0);
-      pixelType_ = format.getStringAsEnum(PropertyKey.PIXEL_TYPE.key(),
-            PixelType.class, null);
-      if (pixelWidth_ * pixelHeight_ * pixelType_.getBytesPerPixel() !=
-            rawPixels_.capacity() * pixelType_.getBytesPerComponent()) {
-         throw new IllegalArgumentException("Image width, height, and pixel type do not match pixel array size");
-      }
+      this(pixels,
+              format.getInteger(PropertyKey.WIDTH.key(), 0),
+              format.getInteger(PropertyKey.HEIGHT.key(), 0),
+              format.getStringAsEnum(PropertyKey.PIXEL_TYPE.key(),
+                      PixelType.class, null).getBytesPerPixel(),
+              format.getStringAsEnum(PropertyKey.PIXEL_TYPE.key(),
+                      PixelType.class, null).getNumberOfComponents(),
+              coords,
+              metadata);
    }
 
    /**
-    * @param pixels Assumed to be a Java array of either bytes or shorts.
+    * Generates a DefaultImage from pixels, minimal image info, and the
+    * supplied coords and metadata.
+    *
+    * Input pixels will be used directly (i.e., they are not copied).
+    *
+    * @param pixels Image pixels.  Should be a Java array of bytes or shorts (not null).
+    * @param coords Coords to be used for this new image (can be null).
+    * @param metadata Metadata to be used this new image (can be null).
+    * @throws IllegalArgumentException
     */
    public DefaultImage(Object pixels, int width, int height, int bytesPerPixel,
          int numComponents, Coords coords, Metadata metadata) 
          throws IllegalArgumentException {
-      metadata_ = (DefaultMetadata) metadata;
-      if (metadata_ == null) {
-         // Don't allow images with null metadata.
-         metadata_ = new DefaultMetadata.Builder().build();
-      }
-      coords_ = coords;
-
-      rawPixels_ = DirectBuffers.bufferFromArray(pixels);
-      if (rawPixels_ == null || rawPixels_.capacity() < width * height) {
-         throw new IllegalArgumentException("Invalid pixel data " + pixels);
-      }
-      pixelWidth_ = width;
-      pixelHeight_ = height;
+      Preconditions.checkNotNull(pixels);
+      metadata_ = metadata == null ? new DefaultMetadata.Builder().build() :
+              (DefaultMetadata) metadata;
+      coords_ = coords == null ? Coordinates.builder().build() : coords;
 
       int bpc;
       if (pixels instanceof byte[]) bpc = 1;
       else if (pixels instanceof short[]) bpc = 2;
       else if (pixels instanceof int[] &&  bytesPerPixel == 4 &&
-            numComponents == 3) {
-         bpc = 1;            
+              numComponents == 3) {
+         bpc = 1;
       } else throw new UnsupportedOperationException("Unsupported pixel data type");
-      pixelType_ = PixelType.valueFor(bytesPerPixel, bpc,
-            numComponents);
+
+      rawPixels_ = BufferTools.wrapArray(pixels, bpc);
+
+      if (rawPixels_ == null || rawPixels_.capacity() < width * height) {
+         throw new IllegalArgumentException("Invalid pixel data " + pixels);
+      }
+      pixelWidth_ = width;
+      pixelHeight_ = height;
+      pixelType_ = PixelType.valueFor(bytesPerPixel, bpc, numComponents);
    }
 
+   /**
+    * Creates a new image object that shares the pixels of the source image
+    * Attaches the provided coords and metadata
+    * @param source Image to be copied
+    * @param coords will be used for this image
+    * @param metadata Metadat for this new image
+    */
    public DefaultImage(Image source, Coords coords, Metadata metadata) {
       metadata_ = (DefaultMetadata) metadata;
       coords_ = coords;
-      if (source instanceof DefaultImage) {
-         // Just copy their Buffer over directly.
-         rawPixels_ = ((DefaultImage) source).getPixelBuffer();
-      }
-      else {
-         rawPixels_ = DirectBuffers.bufferFromArray(source.getRawPixels());
-      }
+
+      int bytesPerComponent = 0;
+      if (source.getRawPixels() instanceof byte[]) { bytesPerComponent = 1; }
+      else if (source.getRawPixels() instanceof short[]) { bytesPerComponent = 2; }
+      rawPixels_ = BufferTools.wrapArray(source.getRawPixels(), bytesPerComponent);
+
       if (rawPixels_.capacity() == 0) {
          throw new IllegalArgumentException("Pixel data has length 0.");
       }
@@ -236,7 +243,9 @@ public final class DefaultImage implements Image {
       else if (rawPixels_ instanceof ShortBuffer) bpc = 2;
       else throw new UnsupportedOperationException("Unsupported pixel data type");
       pixelType_ = PixelType.valueFor(source.getBytesPerPixel(), bpc,
-            source.getNumComponents());
+              source.getNumComponents());
+
+
    }
 
    @Override
@@ -273,7 +282,7 @@ public final class DefaultImage implements Image {
       if (rawPixels_.hasArray()) {
           return rawPixels_.array();
       }
-      return DirectBuffers.arrayFromBuffer(rawPixels_);
+      return BufferTools.arrayFromBuffer(rawPixels_);
    }
 
    @Override
